+++ server/src/model-adapter.test.ts
@@ -2,7 +2,8 @@ import mockfs from 'mock-fs'
 import SinonRoot from 'sinon'
 import { createConnection, WatchDog } from 'vscode-languageserver'
 import { FileChangeType, Logger, ProtocolConnection, PublishDiagnosticsParams } from 'vscode-languageserver-protocol'
-import { BookNode, Bundle } from './model'
+import { BookNode } from './model/book'
+import { Bundle } from './model/bundle'
 import { BundleLoadManager, jobRunner, Job, pageAsTreeObject, URIPair, bookTocAsTreeCollection } from './model-adapter'
 import { first, FS_PATH_HELPER, loadSuccess, makeBundle } from './model.test'
