+++ server/src/model/image.ts
@@ -1,0 +1,6 @@
+import { Fileish } from './fileish'
+
+export class ImageNode extends Fileish {
+  /* istanbul ignore next */
+  public getValidationChecks() { return [] }
+}
